requirements.txt
streamlit==1.32.0
openai==0.28.1
python-dotenv==1.0.1
tinydb==4.8.0 